+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix a bug in `ReorderabledList` that might sometimes make the ordering being wrong.
+  The item order was sometime unexpectedly in "read-only" mode.
+
 ## [v1.2.0] - 2021-06-10
 
 ### Added
@@ -39,7 +44,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [v1.0.2] - 2021-05-11 [YANKED]
 
-### Fix
+### Fixed
 
 - Avoid useless call to `onValueChange` of `CheckboxGroup`. This might have caused React to re-render
   in an infinite loop in some usecases.
@@ -48,7 +53,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [v1.0.1] - 2021-03-11
 
-### Fix
+### Fixed
 
 - Fix a misconfiguration that prevented TS Lib helpers to be used at runtime